+++ client/app/pages/settings/components/GeneralSettings/BeaconConsentSettings.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import Form from "antd/lib/form";
-import HelpTrigger from "@/components/HelpTrigger";
 import Checkbox from "antd/lib/checkbox";
+import Skeleton from "antd/lib/skeleton";
+import HelpTrigger from "@/components/HelpTrigger";
 import DynamicComponent from "@/components/DynamicComponent";
 import { SettingsEditorPropTypes, SettingsEditorDefaultProps } from "../prop-types";
 
 export default function BeaconConsentSettings(props) {
-  const { values, onChange } = props;
+  const { values, onChange, loading } = props;
 
   return (
     <DynamicComponent name="OrganizationSettings.BeaconConsentSettings" {...props}>
@@ -17,12 +18,16 @@ export default function BeaconConsentSettings(props) {
             <HelpTrigger className="m-l-5 m-r-5" type="USAGE_DATA_SHARING" />
           </span>
         }>
-        <Checkbox
-          name="beacon_consent"
-          checked={values.beacon_consent}
-          onChange={e => onChange({ beacon_consent: e.target.checked })}>
-          Help Redash improve by automatically sending anonymous usage data
-        </Checkbox>
+        {loading ? (
+          <Skeleton title={{ width: 300 }} paragraph={false} active />
+        ) : (
+          <Checkbox
+            name="beacon_consent"
+            checked={values.beacon_consent}
+            onChange={e => onChange({ beacon_consent: e.target.checked })}>
+            Help Redash improve by automatically sending anonymous usage data
+          </Checkbox>
+        )}
       </Form.Item>
     </DynamicComponent>
   );
