+++ client/app/pages/settings/components/prop-types.js
@@ -4,10 +4,12 @@ export const SettingsEditorPropTypes = {
   settings: PropTypes.object,
   values: PropTypes.object,
   onChange: PropTypes.func, // (key, value) => void
+  loading: PropTypes.bool,
 };
 
 export const SettingsEditorDefaultProps = {
   settings: {},
   values: {},
   onChange: () => {},
+  loading: false,
 };
